- Updated kotlin version - AndroidManifest.xml requires now the permission to write to the external file storage - Renamed package
