+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="de.voicegym.voicegym">
 
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
 
     <application
         android:allowBackup="true"
